Use agent in the ScheduledTask rather than needlessly using a var.
